requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0,<3.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
imapclient>=3.0.0
openai>=1.0.0
jinja2>=3.0.0
python-multipart>=0.0.5
pytest>=7.0.0
pytest-asyncio>=0.20.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
itsdangerous>=2.0.0

